fix missing library error and change long long type to int

+++ system/utils/global.h
@@ -20,10 +20,10 @@ int _my_rank;
 int _num_workers;
 
 inline long long _cross_worker_msg_num = 0;
-inline vector<vector<long long>> _worker_comm_matrix;
+inline vector<vector<int>> _worker_comm_matrix;
 
 inline void init_comm_matrix() {
-    _worker_comm_matrix.assign(_num_workers, vector<long long>(_num_workers, 0));
+    _worker_comm_matrix.assign(_num_workers, vector<int>(_num_workers, 0));
 }
 
 inline int get_worker_id()
